Move queries order utils to own file
