+++ packages/utils/src/queries/index.ts
@@ -1,2 +1,3 @@
 export * from './utils';
 export * from './filter';
+export * from './order';
